+++ cli/tests/test_git_push.rs
@@ -153,7 +153,7 @@ fn test_git_push_current_bookmark() {
     let output = work_dir.run_jj(["git", "push"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
-    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=origin)..@
+    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=remote)..@
     Nothing changed.
     [EOF]
     ");
@@ -199,7 +199,6 @@ fn test_git_push_no_matching_bookmark() {
     let output = work_dir.run_jj(["git", "push"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
-    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=origin)..@
     Nothing changed.
     [EOF]
     ");
@@ -214,7 +213,6 @@ fn test_git_push_matching_bookmark_unchanged() {
     let output = work_dir.run_jj(["git", "push"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
-    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=origin)..@
     Nothing changed.
     [EOF]
     ");
@@ -260,7 +258,7 @@ fn test_git_push_other_remote_has_bookmark() {
     let output = work_dir.run_jj(["git", "push"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
-    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=origin)..@
+    Warning: No bookmarks found in the default push revset: remote_bookmarks(remote=remote)..@
     Nothing changed.
     [EOF]
     ");
@@ -2193,6 +2191,87 @@ fn test_git_push_rejected_by_remote() {
         [exit status: 1]
         ");
     });
+}
+
+// TODO: Remove with the `git.subprocess` setting.
+#[test]
+fn test_git_push_git2_warning() {
+    let test_env = TestEnvironment::default();
+    set_up(&test_env);
+    let work_dir = test_env.work_dir("local");
+    test_env.add_config("git.subprocess = false");
+    work_dir
+        .run_jj(["describe", "bookmark1", "-m", "modified bookmark1 commit"])
+        .success();
+    let output = work_dir.run_jj(["git", "push", "--all"]);
+    if cfg!(feature = "git2") {
+        insta::assert_snapshot!(output, @r"
+        ------- stderr -------
+        Warning: `git.subprocess = false` will be removed in 0.30; please report any issues you have with the default.
+        Changes to push to origin:
+          Move sideways bookmark bookmark1 from d13ecdbda2a2 to 0f8dc6560f32
+        [EOF]
+        ");
+    } else {
+        insta::assert_snapshot!(output, @r"
+        ------- stderr -------
+        Changes to push to origin:
+          Move sideways bookmark bookmark1 from 9b2e76de3920 to e5ce6d9a0991
+        [EOF]
+        ");
+    }
+}
+
+#[test]
+fn test_git_push_custom_revset() {
+    let test_env = TestEnvironment::default();
+    set_up(&test_env);
+    let work_dir = test_env.work_dir("local");
+    // add a custom revset which simulates ignoring a `private()` revset.
+    test_env.add_config(
+        r#"
+    [revsets]
+    'push(remote)' = "remote_bookmarks(remote=remote)..@ & ~bookmarks(glob:'local/*') & ~subject(glob:'wip:*')"
+    "#,
+    );
+    work_dir
+        .run_jj(["new", "bookmark2", "-m", "commit to be pushed"])
+        .success();
+    work_dir.run_jj(["new", "-m", "wip: stuff"]).success();
+    work_dir
+        .run_jj(["bookmark", "set", "local/stuff", "-r@"])
+        .success();
+    work_dir
+        .run_jj(["new", "-m", "commit which should pushed"])
+        .success();
+    work_dir
+        .run_jj(["new", "-m", "wip: commit which should not be pushed"])
+        .success();
+    //
+    let output = work_dir.run_jj(["log"]);
+    insta::assert_snapshot!(output, @r"
+    @  kmkuslsw test.user@example.com 2001-02-03 08:05:18 ead3bc40
+    │  (empty) wip: commit which should not be pushed
+    ○  kpqxywon test.user@example.com 2001-02-03 08:05:17 048a6554
+    │  (empty) commit which should pushed
+    ○  yostqsxw test.user@example.com 2001-02-03 08:05:15 local/stuff 0712d559
+    │  (empty) wip: stuff
+    ○  vruxwmqv test.user@example.com 2001-02-03 08:05:14 a615282a
+    │  (empty) commit to be pushed
+    ○  zsuskuln test.user@example.com 2001-02-03 08:05:10 bookmark2 38a20473
+    │  (empty) description 2
+    │ ○  qpvuntsm test.user@example.com 2001-02-03 08:05:08 bookmark1 9b2e76de
+    ├─╯  (empty) description 1
+    ◆  zzzzzzzz root() 00000000
+    [EOF]
+    ");
+    // We should try to push the first two commits but not any containing "wip:".
+    let output = work_dir.run_jj(["git", "push"]);
+    insta::assert_snapshot!(output, @r"
+    ------- stderr -------
+    Nothing changed.
+    [EOF]
+    ");
 }
 
 #[must_use]